refactor: pre-fetch animals server-side for formularios AnimalPicker
Replace server action pattern (fetchActiveAnimals via useEffect) with
server component wrapper pattern used in plam-admin/animales. Both
formularios pages now pre-fetch animals via getActiveAnimalsData() and
pass them as initialAnimals prop to client components, eliminating the
FETCH_ANIMALS_ERROR caused by server action + 'use cache' compatibility
issues in production.
