--- src/app/plam-admin/formularios/actions.ts
@@ -1,7 +1,0 @@
-'use server';
-
-import { getActiveAnimalsData } from '@/lib/data/animals';
-
-export async function fetchActiveAnimals() {
-  return getActiveAnimalsData();
-}
